Improve context typing
