+++ client/src/components/workspace/useWorkspaceState.tsx
@@ -1,10 +1,4 @@
-import {
-    Dispatch,
-    ReactNode,
-    SetStateAction,
-    createContext,
-    useState,
-} from 'react';
+import { Dispatch, SetStateAction, useState } from 'react';
 import { PaneData, SplitData, TabsData } from './workspaceData';
 
 type ResizeType = 'horizontal' | 'vertical' | undefined;
@@ -39,40 +33,9 @@ function useWorkspaceState<T>(
     return [views, setViews, addToFocused, { setAddToFocused }];
 }
 
-const TabContentContext = createContext<
-    (value: unknown, onChange: Dispatch<SetStateAction<unknown>>) => ReactNode
->(() => undefined);
-
-const ResizeContext = createContext<Dispatch<SetStateAction<ResizeType>>>(
-    () => {}
-);
-
-const DragContext = createContext<DragContext<unknown>>([
-    [undefined, undefined],
-    () => {},
-]);
-
-type DragContext<T> = [
-    [T, () => void] | [undefined, undefined],
-    Dispatch<SetStateAction<[T, () => void] | [undefined, undefined]>>,
-];
-
-const SetAddToFocusedContext = createContext<Dispatch<() => Dispatch<unknown>>>(
-    () => {}
-);
-
-const CreateTitleContext = createContext<
-    (value: unknown, index: number) => string
->((_, index) => `Tab ${index}`);
-
 export {
     // eslint-disable-next-line react-refresh/only-export-components
     useWorkspaceState,
-    TabContentContext,
-    ResizeContext,
-    DragContext,
-    SetAddToFocusedContext,
-    CreateTitleContext,
     type WorkspaceControls,
     type ResizeType,
 };
